feat:; Makefile形式にプロジェクトに移行
