+++ backend-svc/src/CppBackendSvc/CppBackendSvc/CppBackendSvc.vcxproj.user
@@ -1,11 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <RemoteDebuggerCommandArguments></RemoteDebuggerCommandArguments>
+    <RemoteDebuggerCommandArguments>
+    </RemoteDebuggerCommandArguments>
     <DebuggerFlavor>LinuxDebugger</DebuggerFlavor>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <RemoteDebuggerCommandArguments>/mnt/CppBackendSvc/CppBackendSvc/Test/testdata1.txt</RemoteDebuggerCommandArguments>
     <DebuggerFlavor>LinuxDebugger</DebuggerFlavor>
+    <RemoteDebuggerCommand>/home/neo/projects/CppBackendSvc/CppBackendSvc.out</RemoteDebuggerCommand>
+    <RemoteDebuggerWorkingDirectory>/home/neo/projects/CppBackendSvc/</RemoteDebuggerWorkingDirectory>
   </PropertyGroup>
 </Project>
